// lyxal-cloud/src/types.tsx
export interface CloudProfile {
	username: string;
	name: string;
	default_org: string;
	picture?: string;
	user_hmac?: string;
	enabled_features: string[];
}
// ...
export interface CloudInstanceType {
	slug: string;
	restricted?: boolean;
	display_name: string;
	description: string;
	cpu: number;
	memory: number;
	price_hour: number;
	enabled?: boolean;
	category: string;
	default_storage_size: number;
	max_storage_size: number;
	compute_units: {
		min?: number;
		max?: number;
	};
}

export interface CloudRegion {
	slug: string;
	description: string;
}
// ...
export interface CloudBillingCountry {
	name: string;
	code: string;
}